added the relevant handlers, commands and queries for the CRUD operations on Screen Profiles
note that currently our way to generate IDs for the screen profiles is not ideal\good practice. it is for now in the repository and it works for our POC.

note as well that i did not know how to represent a response after deleting a profile. when we create/update we just return the modified profile
in delete however it didnt feel right to return a deleted screen so i made a response object for delete profile where the contracts and dtos are to represent a response. if you have a better idea to handle this lemme know.

note that we do not work with a db yet, all in repo cache